chore: support data store using pure napi (#287)
Node was one of the earliest bridges we started on. Looking back it, it definitely needs some work. 

We rely on the Statsig InstanceStore rather then simply using napi-rs to manage this for us. We also never hooked it up to a test runner, so the tests that are there are now broken.

In this rewrite, I attempt to use more of napi-rs functionality.
- Use Napi objects/classes with a "inner" ref to a Statsig struct (Free GC).
- Use Napi type conversions wherever possible. This cuts down on JSON.stringifying everything
- Simplify build process. Since we are now only using Napi, we don't need the custom build step and Typescript mapping classes

The goal here is to simplify the bridging logic and have decent test coverage from the outset.