Start sending checksum in DCS request params (#338)
## Reason for change
### Background
To date, we have used the "time" field on Download Config Specs (DCS) to invalidate caches and version payloads. However, we have since discovered that this method of cache invalidation is imperfect and can lead to permanently stale caches.

The problem stems from how we calculate the "time" field; it is the maximal modified time of all of the configs contained in a Statsig project, and it is calculated at the time we generate DCS payloads.

However, at any given time, we cannot guarantee that the service that generates DCS payloads has received all config updates up to a time _X_. It is possible that changes from time _X-1_ and earlier land in our storage _just after_ we generate a DCS. In these cases, the DCS will be generated with `"time": X`, and even if the DCS is re-generated with changes from `X-1` and earlier, the time field will still be `"time": X`. (It is _possible_ to guarantee the order that changes land in storage, but it would come at the cost of propagation latency, which we optimize for.)

Caches that contain the initial DCS with `"time": X` will not be invalidated until the "time" field increments, which won't happen until after another update to the Statsig project.

Likewise, Server SDKs will not receive updated payloads until the "time" field increments.

### About the change
To resolve this issue, we're introducing a "checksum" field to be used in addition to the "time" field to invalidate caches. The checksum represents a hash of the DCS contents. We will invalidate caches if: 
- the new "time" field is greater than the cached "time" field, or
- the "checksum" doesn't match

This works because the service that generates DCS will always produce a DCS that is _at least as up-to-date as_ the last one it generated. If the checksum changes, we can be almost sure that the DCS is newer.