Move response deserialization into networking layer [1/n] (#679)
Introduces a generic `RequestArgs` type: `RequestArgsTyped<T>`, which contains a response deserializer that converts a `Vec<u8>` to `T`

This will callers to pass in a deserializer to `NetworkClient`, which will allow us to move all deserialization/decompression into `NetworkClient`, instead of requiring that each `SpecsAdapter` handle this itself.

This PR does not change existing behavior. It just introduces the new type and makes the minimal changes necessary to keep the compiler happy.