Swap o uing bulk compression for zstd in rust which should be more efficent than our current approach. (#741)
![Screenshot 2025-04-04 at 6.02.50 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/1nlj2EHxzALmiaZzOaCf/aaadab22-b0a1-425f-960e-78aef790b01b.png)

Currently reseting the zstd decoder object is using up a ton of the cpu time for l7, I *think* this is because we are not re-using the object. Reading online the suggestion is to instead use the bulk api or to re-use the obejct. I am going to try the bulk api since we already have all the data we want to encode when we start. 

I should be able to confirm that this is better if we launch it in l7