chore: reuse loggable user json payloads if in memory (#805)
As we have the logic today, we are paying a large perf hit on every exposure event as we need to json stringify the User to lock it as it was at the time of the exposure. This is particularly ugly on repeated checks using the same User object.

In this PR, we move instead to storing the JSON data once per unique user object based on a `full_user_key`. If an identical user is used across checks, we will reuses the previous JSON data created rather then creating a new one.


- Full user key is just a concatination of all user fields + statsig environment
- JSON data is stored by the full user key in the LOGGABLE_USER_STORE lazy static hash map
- When the strong count of a loggable user drops to 1, we remove the entry in the hashmap, ensuring the hash map does not grow indefinitely