chore: allow some waiting during dedupe (#1009)
try_write/try_read are non blocking checks on whether the lock is currently free to use.

If they fail, we effectively just ignore the sampling logic and log the exposure.

I originally did this to make things faster, but the gain was minimal and we may be logging more than we need. 

Switching to the blocking version will ensure we sample correctly