chore: move StatsigUser data to inner UserData arc (#1025)
NOTE: It is known that Kong will fail on this until I update `core-rust/main.rs`


This breaks `StatsigUser` into two parts. An outer and an inner. The outer contains the getter/setters helpers and is passed into every `statsig.rs` function (as it always was).

The inner is a new struct `UserData` that houses all the data and is wrapped in an `Arc`. 

By doing it this way, we can avoid the cost serializing the User during exposure enqueue.

When an exposure occurs, we just clone the `Arc` into the `StatsigUserLoggable`. If the StatsigUser is then updated at some later point, the inner data is cloned into a new `Arc`, leaving all previous references as they were at the time of exposure (See `Arc::make_mut` https://doc.rust-lang.org/std/sync/struct.Arc.html#method.make_mut).