Be less strict about what HTTP status codes should not be retried (#1296)
In the same spirit as #1211, this PR aims to make Server Core a bit more flexible about what HTTP error codes that can be retried. 

The reason for this is that explicitly listing out error codes that can be retried is more error prone than the opposite. 
- When returning new status codes from origin (e.g. `429 Too Many Requests` for rate limiting), we may forget that Server Core will not retry this status code despite that being standard behavior.
- We have already missed a status code that happens frequently in practice (status code 0) and this may happen again.

The status codes that I've listed out as "non-retryable":
- `400 Bad Request` - This means that there was something wrong with the client request and is not considered to be a transient error.
- `403 Forbidden` - We were able to authenticate the caller but we decided that they weren't allowed to access the resource. This is not a transient error.
- `405 Method Not Allowed` - The target resource does not support the HTTP method used. This is not a transient error.
- `501 Not Implemented` - We do not support the HTTP request method used. This is not a transient error.

Notably, I did not include `401 Unauthorized` in this list. My reason is that it is possible that we incorrectly return `401 Unauthorized` upon transient data access failures when fetching SDK Key information.