chore: remove range header for CDN download id list url (#2158)
Remitly reports that they are still seeing 416 after upgrading.

Server side did not respond with 416. It looks like we are still including the range header so that CF will still make a range check even if the cached content is up to date. If the range in the header is smaller than the actual cached file size, CF will throw error.